pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["fuku68 <tofirstplace@gmail.com>"]

[tool.poetry.scripts]
test = "src.slackclient:main"

[tool.poetry.dependencies]
python = "^3.6"
slackbot = "^1.0.0"
boto3 = "^1.17.22"
uvicorn = "^0.13.4"
fastapi = "^0.63.0"
python-dotenv = "^0.15.0"
pydantic = "1.7.3"
slackclient = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^5.7.0"
mypy = "^0.812"
flake8 = "^3.8.4"
autoflake = "^1.4"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
